Fix read depth freeze
The following file was causing the app to freeze:

https://s3.amazonaws.com/iobio/samples/test_files/Horizon-HD728_Panel1385_Hyb11_HFW2KBBXX.bam

The cause was that it was spending a lot of time using the
javascript-natural-sort library to sort chromosome names. This
file in particularlar exposed the problem because it has a lot
of ref names. The solution to this was to use the faster
Intl.Collator class which is built-in to modern browsers.

Another problem was that it was still rendering all the refs,
which took a long time. There was already a warning that was
supposed to pop up in this case, but it was buggy and didn't
actually stop the rendering. This commit also fixes that.