Nest depth arrays in an object
This is in preparation for adding extra information to the
readDepth data. For example, currently mapped and unmapped
are stored directly on the arrays. This works because JavaScript
arrays are also objects, but it isn't idiomatic and could be
confusing. This change will allow these and similar attributes to
be added to a normal object, as siblings to the actual depth
array.